requirements.txt
aiofiles == 24.1.0
aiogram == 3.20.0.post0
aiohappyeyeballs == 2.6.1
aiohttp	== 3.11.18
aiosignal == 1.3.2
annotated-types	== 0.7.0
attrs == 25.3.0
cachetools	== 5.5.2
certifi	== 2025.4.26
charset-normalizer == 3.4.2
frozenlist == 1.6.2
google-api-core	== 2.25.1
google-api-python-client == 2.172.0
google-auth	== 2.40.3
google-auth-httplib2 == 0.2.0
google-auth-oauthlib == 1.2.2
googleapis-common-protos == 1.70.0
gspread	== 6.2.1
httplib2 == 0.22.0
idna ==	3.10
magic-filter == 1.0.12
multidict == 6.4.4
oauthlib == 3.2.2
pip	== 25.1.1
propcache == 0.3.1
proto-plus == 1.26.1
protobuf == 6.31.1
pyasn1 == 0.6.1
pyasn1_modules == 0.4.2
pydantic == 2.11.7
pydantic-settings == 2.9.1
pydantic_core == 2.33.2
pyparsing == 3.2.3
python-dotenv == 1.1.0
requests == 2.32.4
requests-oauthlib == 2.0.0
rsa	== 4.9.1
typing-inspection == 0.4.1
typing_extensions == 4.14.0
uritemplate	== 4.2.0
urllib3	== 2.4.0
yarl == 1.20.0